Renamed API and Queries correctly
