+++ apps/frontend-react/src/api/stream.ts
