+++ apps/frontend-react/src/components/transmission/upload/TransmissionUploadForm.tsx
@@ -4,7 +4,7 @@ import { useForm } from "@mantine/form";
 import { useDisclosure } from "@mantine/hooks";
 import { IconMusic, IconPhoto, IconUpload, IconX } from "@tabler/icons-react";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { apiCreateTransmission } from "../../../api/TransmissionAPI";
+import { apiCreateTransmission } from "../../../api/transmission";
 import isURLSafe from "../../../utils/isURLSafe";
 
 interface FormValues {
